--- types/base_config.go
@@ -1,4 +1,0 @@
-package types
-
-type BaseConfig struct {
-}
